add, sub, mul and neg func proxies on Fp
Fp derives the Add, Sub, Mul and Neg trait and provides an
implementation for the add, sub, mul and neg function. The trait
functions call the underlying add, sub, mul and neg function defined
by Fp. The problem is, these functions are const, which is not allowed
in Trait functions.

Clippy trips over the way add and sub are called on a Fp instance,
since the compiler can automatically borrow it. When the borrow is
removed however, the 'cannot call non-const' error arises. To
circumvent this, I've added proxy functions for add, sub, mul and
neg in the Fp implementation.

Fix immediately dereferenced by compiler warnings

Solve automatic borrow warnings

Add # safety doc annotation to unsafe funcs

+++ src/multiscalar_mul.rs
@@ -122,7 +122,7 @@ fn to_radix_2w_size_hint(w: usize) -> usize {
         6 => (256 + w - 1) / w as usize,
         7 => (256 + w - 1) / w as usize,
         // See comment in to_radix_2w on handling the terminal carry.
-        8 => (256 + w - 1) / w + 1 as usize,
+        8 => (256 + w - 1) / w + 1_usize,
         _ => panic!("invalid radix parameter"),
     };
